fix(comments): only calculate current caret element on demand
